aider: Fix the error `Don't cast a nullable value to a non-nullable type` by adding a not-null assertion to the result of the `signin` method call.
# Aider chat conversation:

USER: Fix the error `Don't cast a nullable value to a non-nullable type.
Try adding a not-null assertion ('!') to make the type non-nullable.`

ASSISTANT: To fix the error `Don't cast a nullable value to a non-nullable type. Try adding a not-null assertion ('!') to make the type non-nullable.`, we need to modify the `lib/src/surrealdb_wasm.dart` file where the `signin` method is implemented. We will add a not-null assertion (`!`) to the result of the `signin` method call to ensure it is treated as a non-nullable `String`.

Here is the *SEARCH/REPLACE block* for the change:

lib/src/surrealdb_wasm.dart
<source>dart
<<<<<<< SEARCH
    final result = await promiseToFuture<Object?>(
      _surreal.signin(jsonEncode(credentials)),
    );
    return dartify(result);
=======
    final result = await promiseToFuture<String>(
      _surreal.signin(jsonEncode(credentials)),
    );
    return dartify(result)!;
>>>>>>> REPLACE
</source>

This change ensures that the result of the `signin` method is cast to a non-nullable `String` by using the not-null assertion (`!`).